Converted to Collection
Converted to collection format 

The latest modules are on GitHub but yet to be packaged up as a collection.  

ansible-galaxy collection install dellemc.powermax-1.1.tar.gz ~/.ansible/collections

Once the modules are installed as a collection a new keyword is needed to import into playbooks   Playbook header would look like this 

---
- name: Provisioning storage for Powermax
  hosts: localhost
  connection: local
  gather_facts: no
  collections:
      - dellemc.powermax

All other tasks are as normal.